lazy load some of the larger components
