+++ lib/components/src/blocks/ArgsTable/ArgValue.tsx
@@ -3,9 +3,9 @@ import { styled } from '@storybook/theming';
 import memoize from 'memoizerific';
 import uniq from 'lodash/uniq';
 import { PropSummaryValue } from './types';
-import { WithTooltipPure } from '../../tooltip/WithTooltip';
+import { WithTooltipPure } from '../../tooltip/lazy-WithTooltip';
 import { Icons } from '../../icon/icon';
-import { SyntaxHighlighter } from '../../syntaxhighlighter/syntaxhighlighter';
+import { SyntaxHighlighter } from '../../syntaxhighlighter/lazy-syntaxhighlighter';
 import { codeCommon } from '../../typography/shared';
 
 interface ArgValueProps {
