--- lib/components/src/controls/index.ts
@@ -1,11 +1,0 @@
-export * from './types';
-
-export * from './Array';
-export * from './Boolean';
-export * from './Color';
-export * from './Date';
-export * from './Number';
-export * from './options';
-export * from './Object';
-export * from './Range';
-export * from './Text';
