+++ lib/components/src/syntaxhighlighter/lazy-syntaxhighlighter.tsx
@@ -1,0 +1,9 @@
+import React, { Suspense } from 'react';
+
+const LazySyntaxHighlighter = React.lazy(() => import('./syntaxhighlighter'));
+
+export const SyntaxHighlighter = (props: React.ComponentProps<typeof LazySyntaxHighlighter>) => (
+  <Suspense fallback={<div />}>
+    <LazySyntaxHighlighter {...props} />
+  </Suspense>
+);
