+++ lib/components/src/syntaxhighlighter/syntaxhighlighter-types.ts
@@ -1,0 +1,20 @@
+import React from 'react';
+
+// @ts-ignore
+export { createElement as createSyntaxHighlighterElement } from 'react-syntax-highlighter';
+
+export interface SyntaxHighlighterRendererProps {
+  rows: any[];
+  stylesheet: string;
+  useInlineStyles: boolean;
+}
+
+export interface SyntaxHighlighterProps {
+  language: string;
+  copyable?: boolean;
+  bordered?: boolean;
+  padded?: boolean;
+  format?: boolean;
+  className?: string;
+  renderer?: (props: SyntaxHighlighterRendererProps) => React.ReactNode;
+}
