+++ lib/components/src/syntaxhighlighter/syntaxhighlighter.tsx
@@ -15,14 +15,11 @@ import tsx from 'react-syntax-highlighter/dist/cjs/languages/prism/tsx';
 import typescript from 'react-syntax-highlighter/dist/cjs/languages/prism/typescript';
 
 import { PrismLight as ReactSyntaxHighlighter } from 'react-syntax-highlighter';
-// @ts-ignore
-import createElement from 'react-syntax-highlighter/dist/cjs/create-element';
 import { ActionBar } from '../ActionBar/ActionBar';
 import { ScrollArea } from '../ScrollArea/ScrollArea';
 
 import { formatter } from './formatter';
-
-export { createElement as createSyntaxHighlighterElement };
+import type { SyntaxHighlighterProps } from './syntaxhighlighter-types';
 
 ReactSyntaxHighlighter.registerLanguage('jsx', jsx);
 ReactSyntaxHighlighter.registerLanguage('json', json);
@@ -92,21 +89,6 @@ const Code = styled.code({
   opacity: 1,
 });
 
-export interface SyntaxHighlighterRendererProps {
-  rows: any[];
-  stylesheet: string;
-  useInlineStyles: boolean;
-}
-export interface SyntaxHighlighterProps {
-  language: string;
-  copyable?: boolean;
-  bordered?: boolean;
-  padded?: boolean;
-  format?: boolean;
-  className?: string;
-  renderer?: (props: SyntaxHighlighterRendererProps) => React.ReactNode;
-}
-
 export interface SyntaxHighlighterState {
   copied: boolean;
 }
@@ -114,6 +96,7 @@ export interface SyntaxHighlighterState {
 type ReactSyntaxHighlighterProps = ComponentProps<typeof ReactSyntaxHighlighter>;
 
 type Props = SyntaxHighlighterProps & ReactSyntaxHighlighterProps;
+
 export const SyntaxHighlighter: FunctionComponent<Props> = ({
   children,
   language = 'jsx',
@@ -168,3 +151,5 @@ export const SyntaxHighlighter: FunctionComponent<Props> = ({
     </Wrapper>
   );
 };
+
+export default SyntaxHighlighter;
